Isolate EH shape classifier tests
Move the EventProvider residual classifier pins out of CompletenessTests so the #1135 curation PRs can merge independently.
